feat(playground): prismify smoke test data, add docs, rename "npm run dev" -> "npm run planetscale"
