+++ query-engine/js-connectors/playground/prisma/schema.prisma
@@ -1,11 +1,12 @@
 generator client {
   provider = "prisma-client-js"
-  previewFeatures = ["jsConnectors"]
+  // previewFeatures = ["jsConnectors"]
 }
 
 datasource db {
-  provider = "@prisma/mysql"
-  url      = env("TEST_DATABASE_URL")
+  provider          = "@prisma/mysql"
+  url               = env("JS_PLANETSCALE_DATABASE_URL")
+  shadowDatabaseUrl = env("JS_PLANETSCALE_SHADOW_DATABASE_URL")
 }
 
 model some_users {
@@ -64,8 +65,6 @@ model type_test {
   blob_null             Bytes?                      @db.Blob
   set_column            String
   set_column_null       String?
-  geometry_column       Unsupported("geometry")
-  geometry_column_null  Unsupported("geometry")?
 }
 
 enum type_test_enum_column {
